test: split PDF visual table crop ownership coverage
